Some tips to clean the up such conditional code
A simple one liner to make the code more readable.  Also make sure the returning statement is the last line of the method so as to be clear what is being returned.  Unfortunately, I cannot effect this code because there are no tests to verify whether my changes will work :(